added a new div for options, when searching for new random people
created a div with class .options
inside it there is a label made using p tag which says "number of results"

there is also a select tag used to make a dropdown menu to select between choice of 10, 20, or 30 results to be shown.

made .options class display:flex so items appear left to right.

centered .options on screen by using combination of position: relative, left: 50% and transform:translate.