store value of name recieved from api into a variable
data.results[i].name.first - to access the first name of the randomly generated person.

data.results[i].name.last - to access the last name of the randomly generated person.

data.results[i] - we use 'i' because it used to to iterate through the for loop

we use +" "+ to combine both strings with a space in the middle